Refactor cloud code to new files and add tests
Add empty preference and storage for cloud so we can check items disappear...
